refactor: extract extension execution helpers
